Change Notifier to an interface.
This is easier for consumers to work with.  The previous behavior of
closing the notifier channel when no more notifications will be sent
has been replaced with a type assertion that the Notifier implements
io.Closer, and calling Close if it does.

This introduces the v2 module due to the breaking change.