Avoid race setting call error
It was possible that both setErr() called by out() and setting a call error by
in() would cause a data race.  Avoid this by acquiring the client's call mutex
whenever a processed call must be finalized.

There are other instances of finalizing the error before the call is ever
added to the client's calls map or touched by in/out.  These don't need to set
the error under the call mutex.  Comment each of these cases.