Panic from finalize() if double finalization is detected
This indicates that a data race likely occurred during concurrent access
setting the call error before finalization.

Just as the Go runtime will crash the process if a race is detected on a map,
likewise we want to do our best to prevent any further process damage with
this panic.