Only finalize mapped all from Go when map entry exists
There was still a double finalization race, detected by the panic introduced
by the previous commit.  If setErr() is called during client teardown, all
currently managed calls will be errored and finalized, and the calls map set
to nil.

Unfortunately, Client.Go was again finalizing the call even after setErr
released the calls mutex, because it had a another variable pointing to it.

To fix this race, look up the call entry in the map and only finalize and
delete it when it present.